Fix remaining compilation errors
- Fix shell_parse to use mutable pointer for string parsing
- Fix utils_strlen to return size_t (matches header)

+++ src/shell/utils.c
@@ -202,8 +202,8 @@ char* utils_strncpy(char* dest, const char* src, size_t n) {
 /*
  * Get string length
  */
-int utils_strlen(const char* s) {
-    int len = 0;
+size_t utils_strlen(const char* s) {
+    size_t len = 0;
     while (*s++) len++;
     return len;
 }
